Small code cleanup (moved one file, deleted unnecessary logs), fixed bug with redirect from /profile on not logged in
